feat(kubernetes): Discover container images and running containers
With this addition the kubernetes provider can now discover container
images and running containers from the connected kubernetes cluster.

A new Daemonset is introduced which provides entrypoints on each
kubernetes node to discover the images and containers running there. The
Kubernetes provider will query each instance of the daemonset to create
full view of all the containers and images that exist on the kubernetes
cluster. The Daemonset supports Docker and ContainerD, other container
runtimes can be added in the future through a defined interface.